chore(material): migrate most components to TypeScript.
