+++ modules/angular2_material/.clang-format
@@ -1,0 +1,12 @@
+Language:        JavaScript
+BasedOnStyle:    Google
+ColumnLimit:     100
+
+TabWidth:                2
+ContinuationIndentWidth: 4
+MaxEmptyLinesToKeep    : 2
+
+AllowShortBlocksOnASingleLine:              false
+AllowShortIfStatementsOnASingleLine:        false
+AllowShortLoopsOnASingleLine:               false
+AllowShortFunctionsOnASingleLine:           Empty
