--- modules/angular2_material/src/components/progress-circular/progress_circular.js
@@ -1,13 +1,0 @@
-import {Component} from 'angular2/src/core/annotations_impl/annotations';
-import {View} from 'angular2/src/core/annotations_impl/view';
-
-@Component({
-  selector: 'md-progress-circular'
-})
-@View({
-  templateUrl: 'angular2_material/src/components/progress-circular/progress_circular.html'
-})
-export class MdProgressCircular {
-  constructor() {
-  }
-}
