+++ modules/angular2_material/src/components/progress-circular/progress_circular.ts
@@ -1,0 +1,7 @@
+import {Component, View} from 'angular2/angular2';
+
+@Component({selector: 'md-progress-circular'})
+@View({templateUrl: 'angular2_material/src/components/progress-circular/progress_circular.html'})
+export class MdProgressCircular {
+  constructor() {}
+}
