+++ modules/angular2_material/src/components/switcher/switch.ts
@@ -1,8 +1,6 @@
-import {Component} from 'angular2/src/core/annotations_impl/annotations';
-import {View} from 'angular2/src/core/annotations_impl/view';
-import {Attribute} from 'angular2/src/core/annotations_impl/di';
+import {Component, View, Attribute} from 'angular2/angular2';
 import {isPresent} from 'angular2/src/facade/lang';
-import {KEY_SPACE} from 'angular2_material/src/core/constants'
+import {KEY_SPACE} from 'angular2_material/src/core/constants';
 import {KeyboardEvent} from 'angular2/src/facade/browser';
 import {NumberWrapper} from 'angular2/src/facade/lang';
 
@@ -10,23 +8,12 @@ import {NumberWrapper} from 'angular2/src/facade/lang';
 
 @Component({
   selector: 'md-switch',
-  properties: {
-    'checked': 'checked',
-    'disabled': 'disabled'
-  },
-  hostListeners: {
-    'keydown': 'onKeydown($event)'
-  },
-  hostProperties: {
-    'checked': 'attr.aria-checked',
-    'disabled_': 'attr.aria-disabled',
-    'role': 'attr.role'
-  }
+  properties: {'checked': 'checked', 'disabled': 'disabled'},
+  hostListeners: {'keydown': 'onKeydown($event)'},
+  hostProperties:
+      {'checked': 'attr.aria-checked', 'disabled_': 'attr.aria-disabled', 'role': 'attr.role'}
 })
-@View({
-  templateUrl: 'angular2_material/src/components/switcher/switch.html',
-  directives: []
-})
+@View({templateUrl: 'angular2_material/src/components/switcher/switch.html', directives: []})
 export class MdSwitch {
   /** Whether this switch is checked. */
   checked: boolean;
@@ -37,7 +24,7 @@ export class MdSwitch {
   tabindex: number;
   role: string;
 
-  constructor(@Attribute('tabindex') tabindex: String) {
+  constructor(@Attribute('tabindex') tabindex: string) {
     this.role = 'checkbox';
     this.checked = false;
     this.tabindex = isPresent(tabindex) ? NumberWrapper.parseInt(tabindex, 10) : 0;
@@ -67,4 +54,3 @@ export class MdSwitch {
     this.checked = !this.checked;
   }
 }
-
