+++ modules/examples/src/material/input/demo_app.html
@@ -34,10 +34,10 @@
     <input disabled value="Firefly">
   </md-input-container>
 
-  <h3>textarea</h3>
+  <!--<h3>textarea</h3>
   <md-input-container>
     <label>What I did on my summer vaccation</label>
     <textarea></textarea>
-  </md-input-container>
+  </md-input-container>-->
 
 </div>
